Update iOS client to match Android client
Security
--------

The IV is no longer hard-coded.

The id="pix" element is no longer required. This removes the only previous traffic indicator that snique was in use.

Any resource referrenced by src= can now carry a snique eTag, not just images.

Any hidden message is sent as a local notification. This improves security if you are only able to surf the web in public but are able to retire to a private location to read your message afterwards. Ensure that in the Notifications section of the settings app you have configured snique to have "Alert Style" set "None" and "View in Lock Screen" set "OFF".

Architecture
------------

Decoding is now performed in SCSSniqueDecoder class. This decouples the decoding logic from the hex-acquisition logic making it simpler to use sources other than HTTP eTags in the future.