General improvements, refactoring
Message now immediately follows the IV rather than beginning at the next eTag boundary. This means that there is no longer any need to insert random hex characters after the IV to pad it to an eTag boundary before beginning the message. This only makes a difference if the number of bytes in a block is not a multiple of the number of bytes encoded in an eTag.

Non-view-controlling code is no longer in the view controller.

The snique decoder has been optimised to reduce its memory usage.

The logic to extract snique messages from a webpage has been moved into its own class.

The logic to extract bytes from a string containing a mix of has and other characters has been moved into an NSData category.